feat: integrate Google Analytics and refactor Home component for improved structure and performance
